3.0draft1 * Deal with a variety of errors reported during assembly, mostly GBIF related, many related to Stramenopiles * Move patches into their own file, so that the data deposit can point to it * Bring in the latest additions * Improve conflict reporting (for writeup example-finding) * Fix small bug in id assignment (range.json processing) * Improve taxon comparator (used to promote deterministic builds)
